fix(stella-tui): unbreak main — split the prompt queue out of deck.rs to get back under its file-size ceiling
PR #1742 added a seven-line cross-store comment to deck.rs, landing it
at 1517 lines against its 1510 baseline ceiling, so the file-size gate
is red on main and every open PR inherits the failure.

Relieve the ratchet the prescribed way (the settlement.rs pattern):
extract a coherent cluster instead of raising the baseline. The
QueuedPrompt/PromptQueue pair moves to deck/prompt_queue.rs, re-exported
from deck so every call site holds; #1742's comment stays. deck.rs lands
at 1474 lines — under the 1500 limit entirely, so its baseline entry is
dropped (the ratchet's own retighten rule) and the god-file tables in
AGENTS.md and the crate README follow the baseline.

Only deck.rs's row leaves the baseline; the other ceilings keep their
existing slack rather than retightening, so in-flight PRs that grew a
file within its recorded ceiling stay green.
