+++ scripts/file-size-baseline.txt
@@ -35,7 +35,6 @@
 1569 crates/stella-tools/src/media.rs
 2305 crates/stella-tools/src/registry.rs
 1839 crates/stella-tools/src/scripts.rs
-1510 crates/stella-tui/src/deck.rs
 1580 crates/stella-tui/src/deck_render.rs
 4028 crates/stella-tui/src/deck_ui.rs
 1665 crates/stella-tui/src/views/engine.rs
